\name{Bivariate.Pareto-package}
\alias{Bivariate.Pareto-package}
\docType{package}
\title{
\packageTitle{Bivariate.Pareto}
}
\description{
\packageDescription{Bivariate.Pareto}
}
%\details{

%}
\author{
\packageAuthor{Bivariate.Pareto}

Maintainer: \packageMaintainer{Bivariate.Pareto}
}
\references{
Shih et al. (2018), Fitting competing risks data to bivariate Pareto models, Communications in Statistics - Theory and Methods, to appear.
}

