% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SN.Pareto.R
\name{SN.Pareto}
\alias{SN.Pareto}
\title{Generate samples from the SNBP distribution}
\usage{
SN.Pareto(n, Alpha0, Alpha1, Alpha2, Gamma)
}
\arguments{
\item{n}{Sample size.}

\item{Alpha0}{Copula parameter \eqn{\alpha_{0}} with restricted range.}

\item{Alpha1}{Positive scale parameter \eqn{\alpha_{1}} for the Pareto margin.}

\item{Alpha2}{Positive scale parameter \eqn{\alpha_{2}} for the Pareto margin.}

\item{Gamma}{Common positive shape parameter \eqn{\gamma} for the Pareto margins.}
}
\value{
\item{X}{\code{X} is asscoiated with the parameters \code{Alpha1} and \code{Gamma}.}
\item{Y}{\code{Y} is asscoiated with the parameters \code{Alpha2} and \code{Gamma}.}
}
\description{
Generate samples from the Sankaran and Nair bivairate Pareto (SNBP) distribution (Sankaran and Nair, 1993).
}
\details{
The admissible range of \code{Alpha0} (\eqn{\alpha_{0}}) is \eqn{0 \leq \alpha_{0} \leq (\gamma+1) \alpha_{1} \alpha_{2}.}
}
\examples{
library(Bivariate.Pareto)
SN.Pareto(5,2,1,1,1)
}
\references{
Sankaran PG, Nair NU (1993), A bivariate Pareto model and its applications to reliability, Naval Research Logistics, 40(7): 1013-1020.

Shih J-H, Lee W, Sun L-H, Emura T (2018), Fitting competing risks data to bivariate Pareto models, Communications in Statistics - Theory and Methods, doi: 10.1080/03610926.2018.1425450.
}
