\name{Bivariate.Pareto-package}
\alias{Bivariate.Pareto-package}
\docType{package}
\title{
\packageTitle{Bivariate.Pareto}
}
\description{
Perform competing risks analysis under bivariate Pareto models. See Shih et al. (2018) for details.
}
%\details{

%}
\author{
\packageAuthor{Bivariate.Pareto}

Maintainer: \packageMaintainer{Bivariate.Pareto}
}
\references{
Shih et al. (2018), Fitting competing risks data to bivariate Pareto models, Communications in Statistics - Theory and Methods, doi: 10.1080/03610926.2018.1425450.
}

