\name{Bivariate.Pareto-package}
\alias{Bivariate.Pareto-package}
\docType{package}
\title{
\packageTitle{Bivariate.Pareto}
}
\description{
Perform competing risks analysis under bivariate Pareto models. See Shih et al. (2018) for details.
}
\details{
The functions in this package are based on latent failure time models with competing risks in Shih et al. (2018). However, they can be adapted to dependent censoring models in Emura and Chen (2018). See \code{MLE.SN.Pareto} for example.
}
\author{
\packageAuthor{Bivariate.Pareto}

Maintainer: \packageMaintainer{Bivariate.Pareto}
}
\references{
Shih J-H, Lee W, Sun L-H, Emura T (2018), Fitting competing risks data to bivariate Pareto models, Communications in Statistics - Theory and Methods, doi: 10.1080/03610926.2018.1425450.

Emura T, Chen Y-H (2018) Analysis of Survival Data with Dependent Censoring, Copula-Based Approaches, JSS Research Series in Statistics, Springer, in press.
}

